Rename "locking" domain "community"
