+++ src/domain/community/entities/__tests__/activity-metadata.builder.ts
@@ -1,5 +1,5 @@
 import { Builder, IBuilder } from '@/__tests__/builder';
-import { ActivityMetadata } from '@/domain/locking/entities/activity-metadata.entity';
+import { ActivityMetadata } from '@/domain/community/entities/activity-metadata.entity';
 import { faker } from '@faker-js/faker';
 
 export function activityMetadataBuilder(): IBuilder<ActivityMetadata> {
