+++ src/domain/community/entities/activity-metadata.entity.ts
