+++ src/domain/community/entities/campaign-rank.entity.ts
