+++ src/domain/community/entities/rank.entity.ts
@@ -1,0 +1,4 @@
+import { RankSchema } from '@/domain/community/entities/schemas/rank.schema';
+import { z } from 'zod';
+
+export type Rank = z.infer<typeof RankSchema>;
