+++ src/domain/community/entities/schemas/locking-event.schema.ts
