+++ src/domain/community/entities/schemas/rank.schema.ts
