--- src/domain/locking/entities/rank.entity.ts
@@ -1,4 +1,0 @@
-import { RankSchema } from '@/domain/locking/entities/schemas/rank.schema';
-import { z } from 'zod';
-
-export type Rank = z.infer<typeof RankSchema>;
