+++ src/routes/community/community.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { LockingDomainModule } from '@/domain/locking/locking.domain.module';
+import { CommunityDomainModule } from '@/domain/community/community.domain.module';
 import { CommunityService } from '@/routes/community/community.service';
 import { CommunityController } from '@/routes/community/community.controller';
 
 @Module({
-  imports: [LockingDomainModule],
+  imports: [CommunityDomainModule],
   providers: [CommunityService],
   controllers: [CommunityController],
 })
