+++ src/routes/locking/entities/campaign.entity.ts
@@ -1,4 +1,4 @@
-import { Campaign as DomainCampaign } from '@/domain/locking/entities/campaign.entity';
+import { Campaign as DomainCampaign } from '@/domain/community/entities/campaign.entity';
 import { ActivityMetadata } from '@/routes/locking/entities/activity-metadata.entity';
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
