+++ src/routes/locking/locking.module.ts
@@ -1,9 +1,7 @@
 import { Module } from '@nestjs/common';
 import { LockingController } from '@/routes/locking/locking.controller';
-import { LockingDomainModule } from '@/domain/locking/locking.domain.module';
 
 @Module({
-  imports: [LockingDomainModule],
   controllers: [LockingController],
 })
 export class LockingModule {}
